move findAssociatedEntityIds to Statement as static method remove unused deleteEntity
